fix: runs function app project in isolated-process mode
