+++ src/Content/Backend/NV.Templates.Backend.Functions/Properties/launchSettings.json
@@ -1,0 +1,9 @@
+{
+  "profiles": {
+    "NV.Templates.Backend.Functions": {
+      "commandName": "Project",
+      "commandLineArgs": "--port 7096",
+      "launchBrowser": false
+    }
+  }
+}
