+++ src/Content/Backend/NV.Templates.Backend.Functions/host.json
@@ -1,11 +1,11 @@
 {
-  "version": "2.0",
-  "logging": {
-    "applicationInsights": {
-      "samplingExcludedTypes": "Request",
-      "samplingSettings": {
-        "isEnabled": true
-      }
+    "version": "2.0",
+    "logging": {
+        "applicationInsights": {
+            "samplingSettings": {
+                "isEnabled": true,
+                "excludedTypes": "Request"
+            }
+        }
     }
-  }
 }
